Refactored to multiple files
